rename some files, add example CURL commands, add DELETE endpoint, in progress [skip ci]
