+++ apps/leaderboard-backend/src/routes/api/guildsRoutes.ts
