+++ apps/leaderboard-backend/src/validation/schema/userSchema.ts
